+++ packages/assets-controllers/src/CurrencyRateController.ts
@@ -1,12 +1,13 @@
 import type {
-  RestrictedMessenger,
   ControllerGetStateAction,
   ControllerStateChangeEvent,
+  StateMetadata,
 } from '@metamask/base-controller';
 import {
   TESTNET_TICKER_SYMBOLS,
   FALL_BACK_VS_CURRENCY,
 } from '@metamask/controller-utils';
+import type { Messenger } from '@metamask/messenger';
 import type { NetworkControllerGetNetworkClientByIdAction } from '@metamask/network-controller';
 import { StaticIntervalPollingController } from '@metamask/polling-controller';
 import { Mutex } from 'async-mutex';
@@ -51,25 +52,23 @@ export type CurrencyRateControllerActions = GetCurrencyRateState;
 
 type AllowedActions = NetworkControllerGetNetworkClientByIdAction;
 
-type CurrencyRateMessenger = RestrictedMessenger<
+export type CurrencyRateMessenger = Messenger<
   typeof name,
   CurrencyRateControllerActions | AllowedActions,
-  CurrencyRateControllerEvents,
-  AllowedActions['type'],
-  never
+  CurrencyRateControllerEvents
 >;
 
-const metadata = {
+const metadata: StateMetadata<CurrencyRateState> = {
   currentCurrency: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: true,
+    includeInDebugSnapshot: true,
     usedInUi: true,
   },
   currencyRates: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: true,
+    includeInDebugSnapshot: true,
     usedInUi: true,
   },
 };
@@ -113,7 +112,7 @@ export class CurrencyRateController extends StaticIntervalPollingController<Curr
    * @param options - Constructor options.
    * @param options.includeUsdRate - Keep track of the USD rate in addition to the current currency rate.
    * @param options.interval - The polling interval, in milliseconds.
-   * @param options.messenger - A reference to the messaging system.
+   * @param options.messenger - A reference to the messenger.
    * @param options.state - Initial state to set on this controller.
    * @param options.useExternalServices - Feature Switch for using external services (default: true)
    * @param options.fetchMultiExchangeRate - Fetches the exchange rate from an external API. This option is primarily meant for use in unit tests.